+++ xenonpy/mdl/mdl.py
@@ -6,7 +6,7 @@
 import os
 import tarfile
 from pathlib import Path
-from typing import List, Union, Tuple
+from typing import List, Union
 
 import pandas as pd
 import requests
@@ -14,6 +14,7 @@ from sklearn.base import BaseEstimator
 from tqdm import tqdm
 
 from xenonpy.utils import TimedMetaClass
+from .base import BaseQuery
 from .descriptor import QueryDescriptors, QueryDescriptorsWith, UpdateDescriptor, CreateDescriptor, ListDescriptors, \
     GetDescriptorDetail
 from .method import QueryMethods, QueryMethodsWith, UpdateMethod, CreateMethod, ListMethods, GetMethodDetail
@@ -25,7 +26,7 @@ from .modelset import QueryModelsets, QueryModelsetsWith, UpdateModelset, Create
 from .property import QueryPropertiesWith, QueryProperties, UpdateProperty, CreateProperty, ListProperties, \
     GetPropertyDetail
 
-__all__ = ['MDL', 'QueryModelsetsWith', 'QueryModelsets', 'QueryModelDetailsWith', 'QueryModelDetails',
+__all__ = ['MDL', 'GetVersion', 'QueryModelsetsWith', 'QueryModelsets', 'QueryModelDetailsWith', 'QueryModelDetails',
            'UpdateModelset', 'UploadModel', 'GetModelsetDetail', 'GetModelDetail', 'GetModelDetails', 'GetModelUrls',
            'GetModelUrl', 'GetTrainingInfo', 'GetSupplementary', 'GetTrainingEnv', 'ListModelsets',
            'ListModelsWithDescriptor', 'ListModelsWithMethod', 'ListModelsWithModelset', 'ListModelsWithProperty',
@@ -33,6 +34,30 @@ __all__ = ['MDL', 'QueryModelsetsWith', 'QueryModelsets', 'QueryModelDetailsWith
            'CreateModelset', 'UpdateProperty', 'QueryDescriptorsWith', 'QueryDescriptors', 'QueryMethodsWith',
            'QueryMethods', 'UpdateDescriptor', 'UpdateMethod', 'ListDescriptors', 'ListMethods', 'GetMethodDetail',
            'GetDescriptorDetail', 'CreateDescriptor', 'CreateMethod']
+
+
+class GetVersion(BaseQuery):
+    queryable = []
+
+    def __init__(self, *, api_key: str = 'anonymous.user.key',
+                 endpoint: str = 'http://xenon.ism.ac.jp/api'):
+        """
+        Access to XenonPy.MDL library.
+
+        Parameters
+        ----------
+        api_key
+            Not implement yet.
+        """
+        super().__init__(variables={}, api_key=api_key, endpoint=endpoint)
+        self._return_json = True
+
+    def gql(self, *query_vars: str):
+        return '''
+            query {
+                getVersion 
+            }
+            '''
 
 
 class MDL(BaseEstimator, metaclass=TimedMetaClass):
@@ -67,6 +92,10 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
         """"""
         self._api_key = k
 
+    @property
+    def version(self):
+        return GetVersion()()
+
     def __call__(self, *query: str,
                  modelset_has: Union[List[str]] = None,
                  property_has: Union[List[str]] = None,
@@ -77,9 +106,9 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
                  transferred: bool = None,
                  deprecated: bool = None,
                  succeed: bool = None,
-                 ):
+                 ) -> Union[QueryModelDetails, QueryModelDetailsWith]:
         """
-        Query models with specific keywords and download to a specific destination
+        Query models with specific keywords.
 
         Parameters
         ----------
@@ -112,8 +141,8 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         Returns
         -------
-        ret: pd.DataFrame
-            A summary of all downloaded models.
+        query
+            Querying object.
         """
 
         if len(query) > 0:
@@ -139,7 +168,7 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
                      describe: dict,
                      training_env: dict = None,
                      training_info: dict = None,
-                     supplementary: dict = None):
+                     supplementary: dict = None) -> UploadModel:
         """
         Upload model to XenonPy.MDL server.
 
@@ -153,6 +182,8 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         Returns
         -------
+        query
+            Querying object.
 
         """
         variables = dict(
@@ -167,7 +198,7 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         return UploadModel(variables, api_key=self.api_key, endpoint=self.endpoint)
 
-    def get_training_info(self, model_id: int):
+    def get_training_info(self, model_id: int) -> GetTrainingInfo:
         """
         Get training information, e.g. ``train_loss``.
 
@@ -178,40 +209,161 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         Returns
         -------
-        info
-            Training information as data frame.
+        query
+            Querying object.
         """
 
         return GetTrainingInfo({'id': model_id}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def get_training_env(self, model_id: int):
+    def get_training_env(self, model_id: int) -> GetTrainingEnv:
+        """
+        Get model's training environments.
+
+        Parameters
+        ----------
+        model_id
+            Model id.
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return GetTrainingEnv({'id': model_id}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def get_supplementary(self, *, model_id: int):
+    def get_supplementary(self, *, model_id: int) -> GetSupplementary:
+        """
+        Get extract information about the model.
+        For regression model this should be *prediction vs observation* data.
+        For classification model, this will be the *confusion matrix*.
+
+        Parameters
+        ----------
+        model_id
+            Model id.
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return GetSupplementary({'id': model_id}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def get_model_url(self, model_id: int):
-        return GetModelUrl({'id': model_id}, api_key=self.api_key, endpoint=self.endpoint)
+    def get_model_urls(self, *model_ids: int) -> Union[GetModelUrl, GetModelUrls]:
+        """
+        Get download url by model id.
 
-    def get_model_urls(self, model_ids: List[int]):
-        return GetModelUrls({'ids': model_ids}, api_key=self.api_key, endpoint=self.endpoint)
+        Parameters
+        ----------
+        model_ids
+            Model id.
 
-    def get_model_detail(self, model_id: int):
+        Returns
+        -------
+        query
+            Querying object.
+        """
+        if len(model_ids) == 0:
+            raise RuntimeError('input is not non-able')
+        if len(model_ids) == 1:
+            return GetModelUrl({'id': model_ids[0]}, api_key=self.api_key, endpoint=self.endpoint)
+        else:
+            return GetModelUrls({'ids': model_ids}, api_key=self.api_key, endpoint=self.endpoint)
+
+    def get_model_detail(self, model_id: int) -> GetModelDetail:
+        """
+        Get model detail by model id.
+
+        Parameters
+        ----------
+        model_id
+            Model id.
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return GetModelDetail({'id': model_id}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def get_model_details(self, model_ids: List[int]):
+    def get_model_details(self, model_ids: List[int]) -> GetModelDetails:
+        """
+        Get multiply model detail by their ids in one querying.
+
+        Parameters
+        ----------
+        model_ids
+            Model id.
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return GetModelDetails({'ids': model_ids}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def list_models_with_property(self, name: str):
+    def list_models_with_property(self, name: str) -> ListModelsWithProperty:
+        """
+        List all models relevant with given property.
+
+        Parameters
+        ----------
+        name
+            Property name
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return ListModelsWithProperty({'name': name}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def list_models_with_modelset(self, name: str):
+    def list_models_with_modelset(self, name: str) -> ListModelsWithModelset:
+        """
+        List all models relevant with given modelset.
+
+        Parameters
+        ----------
+        name
+            Modelset name
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return ListModelsWithModelset({'name': name}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def list_models_with_method(self, name: str):
+    def list_models_with_method(self, name: str) -> ListModelsWithMethod:
+        """
+        List all models relevant with given method.
+
+        Parameters
+        ----------
+        name
+            Method name
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return ListModelsWithMethod({'name': name}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def list_models_with_descriptor(self, name: str):
+    def list_models_with_descriptor(self, name: str) -> ListModelsWithDescriptor:
+        """
+        List all models relevant with given descriptor.
+
+        Parameters
+        ----------
+        name
+            Descriptor name
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return ListModelsWithDescriptor({'name': name}, api_key=self.api_key, endpoint=self.endpoint)
 
     def query_modelsets(self, query: str = None, *,
@@ -220,9 +372,9 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
                         describe_has: Union[List[str]] = None,
                         private: bool = None,
                         deprecated: bool = None,
-                        ):
+                        ) -> Union[QueryModelsets, QueryModelsetsWith]:
         """
-        Query models with specific keywords and download to a specific destination
+        Query modelsets with specific keywords.
 
         Parameters
         ----------
@@ -248,8 +400,8 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         Returns
         -------
-        ret: pd.DataFrame
-            Matched modelsets.
+        query
+            Querying object.
         """
 
         if query is not None:
@@ -274,9 +426,9 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
                         tags: List[str] = None,
                         private: bool = None,
                         deprecated: bool = None
-                        ):
+                        ) -> UpdateModelset:
         """
-        Upload model to XenonPy.MDL server.
+        Update modelset information by modelset id.
 
         Parameters
         ----------
@@ -290,7 +442,8 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         Returns
         -------
-
+        query
+            Querying object.
         """
         with_ = dict(
             name=name,
@@ -312,7 +465,7 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
                        private: bool = False
                        ):
         """
-        Create modelset..
+        Create modelset.
 
         Parameters
         ----------
@@ -324,7 +477,8 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         Returns
         -------
-
+        query
+            Querying object.
         """
         with_ = dict(
             name=name,
@@ -337,19 +491,39 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         return CreateModelset({'with_': with_}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def list_modelsets(self):
+    def list_modelsets(self) -> ListModelsets:
+        """
+        List all modelsets.
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return ListModelsets(api_key=self.api_key, endpoint=self.endpoint)
 
-    def get_modelset_detail(self, modelset_id: int):
+    def get_modelset_detail(self, modelset_id: int) -> GetModelsetDetail:
+        """
+        Get modelset detail by modelset id.
+
+        Parameters
+        ----------
+        modelset_id
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return GetModelsetDetail({'id': modelset_id}, api_key=self.api_key, endpoint=self.endpoint)
 
     def query_descriptors(self, query: str = None, *,
                           name_has: Union[List[str]] = None,
                           fullname_has: Union[List[str]] = None,
                           describe_has: Union[List[str]] = None,
-                          ):
+                          ) -> Union[QueryDescriptors, QueryDescriptorsWith]:
         """
-        Query models with specific keywords and download to a specific destination
+        Query descriptors with specific keywords.
 
         Parameters
         ----------
@@ -360,14 +534,14 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
             The part of a model set's name.
             For example, ``modelset_has='test`` will hit ``*test*``
         fullname_has
-            A part of the name of descriptor.
+            Part of the full name.
         describe_has
-            A part of the name of descriptor.
+            Part of the descriptor.
 
         Returns
         -------
-        ret: pd.DataFrame
-            Matched modelsets.
+        query
+            Querying object.
         """
 
         if query is not None:
@@ -387,9 +561,9 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
                           new_name: str = None,
                           describe: str = None,
                           fullname: str = None,
-                          ):
+                          ) -> UpdateDescriptor:
         """
-        Upload model to XenonPy.MDL server.
+        Update descriptor information by name.
 
         Parameters
         ----------
@@ -400,8 +574,10 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         Returns
         -------
-
+        query
+            Querying object.
         """
+
         with_ = dict(
             name=new_name,
             describe=describe,
@@ -415,9 +591,9 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
                          name: str,
                          describe: str = None,
                          fullname: str = None,
-                         ):
+                         ) -> CreateDescriptor:
         """
-        Create modelset..
+        Create descriptor.
 
         Parameters
         ----------
@@ -427,8 +603,10 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         Returns
         -------
-
+        query
+            Querying object.
         """
+
         with_ = dict(
             name=name,
             describe=describe,
@@ -438,19 +616,39 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         return CreateDescriptor({'with_': with_}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def list_descriptors(self):
+    def list_descriptors(self) -> ListDescriptors:
+        """
+        List all descriptor.
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return ListDescriptors(api_key=self.api_key, endpoint=self.endpoint)
 
-    def get_descriptor_detail(self, name: str):
+    def get_descriptor_detail(self, name: str) -> GetDescriptorDetail:
+        """
+        Get descriptor detail by name.
+
+        Parameters
+        ----------
+        name
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return GetDescriptorDetail({'name': name}, api_key=self.api_key, endpoint=self.endpoint)
 
     def query_methods(self, query: str = None, *,
                       name_has: Union[List[str]] = None,
                       fullname_has: Union[List[str]] = None,
                       describe_has: Union[List[str]] = None,
-                      ):
+                      ) -> Union[QueryMethods, QueryMethodsWith]:
         """
-        Query models with specific keywords and download to a specific destination
+        Query methods with specific keywords.
 
         Parameters
         ----------
@@ -461,14 +659,14 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
             The part of a model set's name.
             For example, ``modelset_has='test`` will hit ``*test*``
         fullname_has
-            A part of the name of method.
+            Part of the full name.
         describe_has
-            A part of the name of descriptor.
+            Part of the descriptor.
 
         Returns
         -------
-        ret: pd.DataFrame
-            Matched modelsets.
+        query
+            Querying object.
         """
 
         if query is not None:
@@ -488,9 +686,9 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
                       new_name: str = None,
                       describe: str = None,
                       fullname: str = None,
-                      ):
+                      ) -> UpdateMethod:
         """
-        Upload model to XenonPy.MDL server.
+        Update method information by name.
 
         Parameters
         ----------
@@ -501,8 +699,10 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         Returns
         -------
-
+        query
+            Querying object.
         """
+
         with_ = dict(
             name=new_name,
             describe=describe,
@@ -516,9 +716,9 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
                      name: str,
                      describe: str = None,
                      fullname: str = None,
-                     ):
+                     ) -> CreateMethod:
         """
-        Create modelset..
+        Create method.
 
         Parameters
         ----------
@@ -528,8 +728,10 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         Returns
         -------
-
+        query
+            Querying object.
         """
+
         with_ = dict(
             name=name,
             describe=describe,
@@ -539,10 +741,30 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         return CreateMethod({'with_': with_}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def list_methods(self):
+    def list_methods(self) -> ListMethods:
+        """
+        List all methods.
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return ListMethods(api_key=self.api_key, endpoint=self.endpoint)
 
-    def get_method_detail(self, name: str):
+    def get_method_detail(self, name: str) -> GetMethodDetail:
+        """
+        Get method detail by name.
+
+        Parameters
+        ----------
+        name
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return GetMethodDetail({'name': name}, api_key=self.api_key, endpoint=self.endpoint)
 
     def query_properties(self, query: str = None, *,
@@ -551,9 +773,9 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
                          describe_has: Union[List[str]] = None,
                          symbol_has: Union[List[str]] = None,
                          unit_has: Union[List[str]] = None,
-                         ):
+                         ) -> Union[QueryProperties, QueryPropertiesWith]:
         """
-        Query models with specific keywords and download to a specific destination
+        Query properties with specific keywords.
 
         Parameters
         ----------
@@ -564,20 +786,18 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
             The part of a model set's name.
             For example, ``modelset_has='test`` will hit ``*test*``
         fullname_has
-            A part of the name of methods.
+            Part of the full name.
         describe_has
-            A part of the name of descriptor.
+            Part of the descriptor.
         symbol_has
-            If``True``, searching in regression models,
-            else, searching in classification models.
-            Default is ``True``.
+            Part of the symbol.
         unit_has
-            Model with this mark is deprecated.
+            Part of the unit.
 
         Returns
         -------
-        ret: pd.DataFrame
-            Matched modelsets.
+        query
+            Querying object.
         """
 
         if query is not None:
@@ -601,9 +821,9 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
                         fullname: str = None,
                         symbol: str = None,
                         unit: str = None,
-                        ):
+                        ) -> UpdateProperty:
         """
-        Upload model to XenonPy.MDL server.
+        Update property information by name.
 
         Parameters
         ----------
@@ -616,7 +836,8 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         Returns
         -------
-
+        query
+            Querying object.
         """
         with_ = dict(
             name=new_name,
@@ -635,9 +856,9 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
                        fullname: str = None,
                        symbol: str = None,
                        unit: str = None,
-                       ):
+                       ) -> CreateProperty:
         """
-        Create modelset..
+        Create property.
 
         Parameters
         ----------
@@ -649,8 +870,10 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         Returns
         -------
-
+        query
+            Querying object.
         """
+
         with_ = dict(
             name=name,
             describe=describe,
@@ -662,19 +885,41 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
 
         return CreateProperty({'with_': with_}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def list_properties(self):
+    def list_properties(self) -> ListProperties:
+        """
+        List all properties.
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return ListProperties(api_key=self.api_key, endpoint=self.endpoint)
 
     def get_property_detail(self, name: str):
+        """
+        Get property detail by name.
+
+        Parameters
+        ----------
+        name
+
+        Returns
+        -------
+        query
+            Querying object.
+        """
         return GetPropertyDetail({'name': name}, api_key=self.api_key, endpoint=self.endpoint)
 
-    def pull(self, model_ids: Union[Tuple[int], List[int], pd.Series, pd.DataFrame], save_to: str = '.'):
+    def pull(self, *model_ids: Union[int, pd.Series, pd.DataFrame],
+             save_to: str = '.') -> pd.DataFrame:
         """
+        Download model(s) from XenonPy.MDL server.
 
         Parameters
         ----------
         model_ids
-            List of model ids.
+            Model ids.
             It can be given by a dataframe.
             In this case, the column with name ``id`` will be used.
         save_to
@@ -684,13 +929,16 @@ class MDL(BaseEstimator, metaclass=TimedMetaClass):
         -------
 
         """
-        if isinstance(model_ids, pd.Series):
-            model_ids = model_ids.tolist()
+        if len(model_ids) == 0:
+            raise RuntimeError('input can not be empty')
+        if len(model_ids) == 1:
+            if isinstance(model_ids[0], pd.Series):
+                model_ids = model_ids[0].tolist()
 
-        if isinstance(model_ids, pd.DataFrame):
-            model_ids = model_ids['id'].tolist()
+            if isinstance(model_ids[0], pd.DataFrame):
+                model_ids = model_ids[0]['id'].tolist()
 
-        ret = self.get_model_urls(model_ids)('id', 'url')
+        ret = self.get_model_urls(*model_ids)('id', 'url')
         urls = ret['url'].tolist()
 
         path_list = []