+++ xenonpy/model/utils/__init__.py
@@ -3,4 +3,3 @@
 #  license that can be found in the LICENSE file.
 
 from .metrics import *
-from .parameter_gen import *